requirements.txt
streamlit
pdfplumber
supabase
openai>=1.0.0
python-dotenv
tiktoken
